+++ src/Model/Matera.java
@@ -1,22 +1,15 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
+
 package Model;
 
-/**
- *
- * @author 57301
- */
 public class Matera {
     //Atributos
     private int codigo;
     private String nombre;
     private String material;
     private float tamaño;
-    private float precio;
+    private int precio;
 
-    public Matera(int codigo, String nombre, String material, float tamaño, float precio) {
+    public Matera(int codigo, String nombre, String material, float tamaño, int precio) {
         this.codigo = codigo;
         this.nombre = nombre;
         this.material = material;
@@ -64,11 +57,11 @@ public class Matera {
         this.tamaño = tamaño;
     }
 
-    public float getPrecio() {
+    public int getPrecio() {
         return precio;
     }
 
-    public void setPrecio(float precio) {
+    public void setPrecio(int precio) {
         this.precio = precio;
     }
